requirements.txt

pytesseract~=0.3.13
PyMuPDF~=1.24.11
requests~=2.32.3
fastapi~=0.115.0
uvicorn

pillow~=10.4.0